+++ src/neptune/internal/utils/git.py
@@ -134,7 +134,7 @@ def get_diff(repo: "git.Repo", commit_ref: str) -> Optional[str]:
         from git.exc import GitCommandError
 
         try:
-            return repo.git.diff(commit_ref)
+            return repo.git.diff(commit_ref, index=False)
         except GitCommandError:
             return
     except ImportError:
@@ -183,16 +183,13 @@ def get_upstream_index_sha(repo: "git.Repo") -> Optional[str]:
 
 
 def get_uncommitted_changes(repo: Optional["git.Repo"]) -> Optional[UncommittedChanges]:
-    if not repo.is_dirty(index=False):
-        return
-
     head_index_diff = get_diff(repo, repo.head.name)
 
     upstream_sha = get_upstream_index_sha(repo)
 
-    upstream_index_diff = get_diff(repo, upstream_sha)
-
-    return UncommittedChanges(head_index_diff, upstream_index_diff, upstream_sha)
+    upstream_index_diff = get_diff(repo, upstream_sha) if upstream_sha else None
+    if head_index_diff or upstream_sha or upstream_index_diff:
+        return UncommittedChanges(head_index_diff, upstream_index_diff, upstream_sha)
 
 
 def track_uncommitted_changes(git_ref: Union[GitRef, GitRefDisabled], run: "Run") -> None: