+++ tests/unit/neptune/new/internal/backends/test_hosted_neptune_backend.py
@@ -16,6 +16,7 @@
 import socket
 import unittest
 import uuid
+from pathlib import Path
 from unittest.mock import call
 
 from bravado.exception import (
@@ -87,7 +88,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
         create_artifacts_client.cache_clear()
 
         self.container_types = [ContainerType.RUN, ContainerType.PROJECT]
-        self.dummy_operation_storage = OperationStorage("./tests/dummy_storage")
+        self.dummy_operation_storage = OperationStorage(Path("./tests/dummy_storage"))
 
     @patch("neptune.internal.backends.hosted_neptune_backend.upload_file_attribute")
     @patch("socket.gethostbyname", MagicMock(return_value="1.1.1.1"))