+++ tests/unit/neptune/new/internal/utils/test_dependency_tracking.py
@@ -1,3 +1,4 @@
+import itertools
 import unittest
 from unittest.mock import (
     MagicMock,
@@ -11,27 +12,32 @@ from neptune.internal.utils.dependency_tracking import (
 
 
 class TestDependencyTracking(unittest.TestCase):
-    @patch("subprocess.check_output", return_value=b"some_dependency==1.0.0\n")
+    @patch("neptune.internal.utils.dependency_tracking.distributions")
     @patch("neptune.types.File.from_content")
-    def test_infer_calls_upload_correctly(self, mock_from_content, mock_check_output):
+    def test_infer_calls_upload_correctly(self, mock_from_content, mock_distributions):
+        single_dist = MagicMock()
+        single_dist.metadata = {"Name": "some_dependency", "Version": "1.0.0"}
+        mock_distributions.return_value = itertools.chain([single_dist])
         InferDependenciesStrategy().log_dependencies(run=MagicMock())
 
-        mock_check_output.assert_called_once()
-        mock_from_content.assert_called_once_with("some_dependency==1.0.0\n")
+        mock_distributions.assert_called_once()
+        mock_from_content.assert_called_once_with("some_dependency==1.0.0")
 
-    @patch("subprocess.check_output", return_value=b"")
+    @patch("neptune.internal.utils.dependency_tracking.distributions", return_value=[])
     @patch("neptune.types.File.from_content")
-    def test_infer_does_not_upload_empty_dependency_string(self, mock_from_content, mock_check_output):
+    def test_infer_does_not_upload_empty_dependency_string(self, mock_from_content, mock_distributions):
         InferDependenciesStrategy().log_dependencies(run=MagicMock())
 
-        mock_check_output.assert_called_once()
+        mock_distributions.assert_called_once()
         mock_from_content.assert_not_called()
 
-    @patch("neptune.handler.Handler.upload_files")
-    def test_file_strategy_not_uploading_if_path_incorrect(self, mock_upload_files):
+    @patch("neptune.handler.Handler.upload")
+    @patch("neptune.internal.utils.dependency_tracking.logger")
+    def test_file_strategy_path_incorrect(self, mock_logger, mock_upload):
         FileDependenciesStrategy(path="non-existent_file_path.txt").log_dependencies(run=MagicMock())
 
-        mock_upload_files.assert_not_called()
+        mock_upload.assert_not_called()
+        mock_logger.error.assert_called_once()
 
     @patch("os.path.isfile", return_value=True)
     def test_file_strategy_uploads_correct_path(self, mock_is_file):
@@ -39,8 +45,8 @@ class TestDependencyTracking(unittest.TestCase):
         handler = MagicMock()
         run.__getitem__ = MagicMock()
         run.__getitem__.return_value = handler
-        handler.upload_files = MagicMock()
+        handler.upload = MagicMock()
 
         FileDependenciesStrategy(path="valid_file_path.txt").log_dependencies(run=run)
 
-        handler.upload_files.assert_called_once_with("valid_file_path.txt")
+        handler.upload.assert_called_once_with("valid_file_path.txt")