+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-## [UNRELEASED] neptune 1.3.2
+## [UNRELEASED] neptune 1.3.3
+
+### Fixes
+- Fixed operation processor bug if current working directory is different from the script directory ([#1391](https://github.com/neptune-ai/neptune-client/pull/1391))
+
+##  neptune 1.3.3rc0
+
+### Changes
+- Dependency tracking feature will log an error if a given file path doesn't exist ([#1389](https://github.com/neptune-ai/neptune-client/pull/1389))
+- Use `importlib` instead of `pip freeze` in dependency tracking ([#1389](https://github.com/neptune-ai/neptune-client/pull/1389))
+- Log both uploaded and inferred requirements to the same namespace ([#1389](https://github.com/neptune-ai/neptune-client/pull/1389))
+
+### Fixes
+- Fixed operation processor bug if current working directory is different from the script directory ([#1391](https://github.com/neptune-ai/neptune-client/pull/1391))
+- Tracking uncommitted changes and dependencies will be skipped in case of any exception, to not disturb the run initialization ([#1395](https://github.com/neptune-ai/neptune-client/pull/1395))
+
+## neptune 1.3.2
 
 ### Features
 - Added support for `mlflow` integration ([#1381](https://github.com/neptune-ai/neptune-client/pull/1381))
@@ -7,6 +23,10 @@
 - Fixed GitPython `is_dirty` failing on Windows ([#1371](https://github.com/neptune-ai/neptune-client/pull/1371))
 - Fix SSL errors after forking process ([#1353](https://github.com/neptune-ai/neptune-client/pull/1353))
 - Fixed support of stringify value in series attributes with step ([#1373](https://github.com/neptune-ai/neptune-client/pull/1373))
+- `dict`s and `Namespace`s that are written to runs and contain an empty string "" key now produce a warning and drop
+  the entry with such a key instead of raising an
+  exception ([#1374](https://github.com/neptune-ai/neptune-client/pull/1374))
+- Fix dependency tracking by replacing `pipreqs` with `pip freeze` ([#1384](https://github.com/neptune-ai/neptune-client/pull/1384))
 
 ### Changes
 - Added support of writing to archived project exception ([#1355](https://github.com/neptune-ai/neptune-client/pull/1355))